Moves guest/out of stock scenarios into dedicated file
